+++ TeamProject/WebContent/WEB-INF/views/managePage/managePage.jsp
@@ -27,6 +27,12 @@
 		<input type="submit" value="회원관리">
 		<input type="submit" formaction="bbsList.do" value="공지사항관리">
 		 </form>
+		<br> <br>
+		<button onclick="location.href='itemWriteGo.do'">제품등록</button>
+		<br> <br>
+		<button onclick="location.href='refundManage.do?pageSize=10&pageNum=1&category=userid&keyword='">취소환불 관리</button>
+		<br> <br>
+		<button onclick="location.href='orderManage.do?pageSize=10&pageNum=1&category=userid&keyword='">배송상태 관리</button>
 	</div>
 	<br>
 	<br>
